Create bundle processor
