+++ components/director/internal/open_resource_discovery/processors/common.go
@@ -2,7 +2,9 @@ package processors
 
 import (
 	"context"
+	"errors"
 	"github.com/kyma-incubator/compass/components/director/pkg/log"
+	"github.com/mitchellh/hashstructure/v2"
 )
 
 func addFieldToLogger(ctx context.Context, fieldName, fieldValue string) context.Context {
@@ -18,4 +20,18 @@ func searchInSlice(length int, f func(i int) bool) (int, bool) {
 		}
 	}
 	return -1, false
+}
+
+func equalStrings(first, second *string) bool {
+	return first != nil && second != nil && *first == *second
+}
+
+// HashObject hashes the given object
+func HashObject(obj interface{}) (uint64, error) {
+	hash, err := hashstructure.Hash(obj, hashstructure.FormatV2, &hashstructure.HashOptions{SlicesAsSets: true})
+	if err != nil {
+		return 0, errors.New("failed to hash the given object")
+	}
+
+	return hash, nil
 }
