+++ components/director/internal/open_resource_discovery/service.go
@@ -2,9 +2,7 @@ package ord
 
 import (
 	"context"
-	"encoding/json"
 	"github.com/kyma-incubator/compass/components/director/internal/open_resource_discovery/processors"
-	"strconv"
 	"strings"
 	"sync"
 	"time"
@@ -15,9 +13,6 @@ import (
 
 	"github.com/imdario/mergo"
 	"github.com/kyma-incubator/compass/components/director/pkg/apperrors"
-
-	"github.com/kyma-incubator/compass/components/director/pkg/accessstrategy"
-	"github.com/kyma-incubator/compass/components/director/pkg/graphql"
 
 	"github.com/kyma-incubator/compass/components/director/internal/metrics"
 	directorresource "github.com/kyma-incubator/compass/components/director/pkg/resource"
@@ -36,24 +31,13 @@ import (
 const (
 	// MultiErrorSeparator represents the separator for splitting multi error into slice of validation errors
 	MultiErrorSeparator string = "* "
-	// TenantMappingCustomTypeIdentifier represents an identifier for tenant mapping webhooks in Credential exchange strategies
-	TenantMappingCustomTypeIdentifier = "sap.ucl:tenant-mapping"
-
-	customTypeProperty  = "customType"
-	callbackURLProperty = "callbackUrl"
 )
 
 // ServiceConfig contains configuration for the ORD aggregator service
 type ServiceConfig struct {
 	maxParallelSpecificationProcessors int
 
-	credentialExchangeStrategyTenantMappings map[string]CredentialExchangeStrategyTenantMapping
-}
-
-// CredentialExchangeStrategyTenantMapping contains tenant mappings configuration
-type CredentialExchangeStrategyTenantMapping struct {
-	Mode    model.WebhookMode
-	Version string
+	credentialExchangeStrategyTenantMappings map[string]processors.CredentialExchangeStrategyTenantMapping
 }
 
 type ordFetchRequest struct {
@@ -75,7 +59,7 @@ type MetricsConfig struct {
 }
 
 // NewServiceConfig creates new ServiceConfig from the supplied parameters
-func NewServiceConfig(maxParallelSpecificationProcessors int, credentialExchangeStrategyTenantMappings map[string]CredentialExchangeStrategyTenantMapping) ServiceConfig {
+func NewServiceConfig(maxParallelSpecificationProcessors int, credentialExchangeStrategyTenantMappings map[string]processors.CredentialExchangeStrategyTenantMapping) ServiceConfig {
 	return ServiceConfig{
 		maxParallelSpecificationProcessors:       maxParallelSpecificationProcessors,
 		credentialExchangeStrategyTenantMappings: credentialExchangeStrategyTenantMappings,
@@ -90,8 +74,9 @@ type Service struct {
 	transact persistence.Transactioner
 
 	appSvc                ApplicationService
-	webhookSvc            WebhookService
-	bundleSvc             BundleService
+	webhookSvc            processors.WebhookService
+	bundleSvc             processors.BundleService
+	bundleProcessor       processors.BundleProcessor
 	bundleReferenceSvc    BundleReferenceService
 	apiSvc                APIService
 	eventSvc              EventService
@@ -112,14 +97,12 @@ type Service struct {
 
 	ordWebhookMapping []application.ORDWebhookMapping
 
-	webhookConverter WebhookConverter
-
 	globalRegistrySvc GlobalRegistryService
 	ordClient         Client
 }
 
 // NewAggregatorService returns a new object responsible for service-layer ORD operations.
-func NewAggregatorService(config ServiceConfig, metricsCfg MetricsConfig, transact persistence.Transactioner, appSvc ApplicationService, webhookSvc WebhookService, bundleSvc BundleService, bundleReferenceSvc BundleReferenceService, apiSvc APIService, eventSvc EventService, specSvc SpecService, fetchReqSvc FetchRequestService, packageSvc processors.PackageService, packageProcessor processors.PackageProcessor, productSvc processors.ProductService, productProcessor processors.ProductProcessor, vendorSvc processors.VendorService, vendorProcessor processors.VendorProcessor, tombstoneProcessor TombstoneProcessor, tenantSvc TenantService, globalRegistrySvc GlobalRegistryService, client Client, webhookConverter WebhookConverter, appTemplateVersionSvc ApplicationTemplateVersionService, appTemplateSvc ApplicationTemplateService, labelService LabelService, ordWebhookMapping []application.ORDWebhookMapping, opSvc operationsmanager.OperationService) *Service {
+func NewAggregatorService(config ServiceConfig, metricsCfg MetricsConfig, transact persistence.Transactioner, appSvc ApplicationService, webhookSvc processors.WebhookService, bundleSvc processors.BundleService, bundleReferenceSvc BundleReferenceService, apiSvc APIService, eventSvc EventService, specSvc SpecService, fetchReqSvc FetchRequestService, packageSvc processors.PackageService, packageProcessor processors.PackageProcessor, productSvc processors.ProductService, productProcessor processors.ProductProcessor, vendorSvc processors.VendorService, vendorProcessor processors.VendorProcessor, tombstoneProcessor TombstoneProcessor, tenantSvc TenantService, globalRegistrySvc GlobalRegistryService, client Client, appTemplateVersionSvc ApplicationTemplateVersionService, appTemplateSvc ApplicationTemplateService, labelService LabelService, ordWebhookMapping []application.ORDWebhookMapping, opSvc operationsmanager.OperationService, bundleProcessor processors.BundleProcessor) *Service {
 	return &Service{
 		config:                config,
 		metricsCfg:            metricsCfg,
@@ -127,6 +110,7 @@ func NewAggregatorService(config ServiceConfig, metricsCfg MetricsConfig, transa
 		appSvc:                appSvc,
 		webhookSvc:            webhookSvc,
 		bundleSvc:             bundleSvc,
+		bundleProcessor:       bundleProcessor,
 		bundleReferenceSvc:    bundleReferenceSvc,
 		apiSvc:                apiSvc,
 		eventSvc:              eventSvc,
@@ -142,7 +126,6 @@ func NewAggregatorService(config ServiceConfig, metricsCfg MetricsConfig, transa
 		tenantSvc:             tenantSvc,
 		globalRegistrySvc:     globalRegistrySvc,
 		ordClient:             client,
-		webhookConverter:      webhookConverter,
 		appTemplateVersionSvc: appTemplateVersionSvc,
 		appTemplateSvc:        appTemplateSvc,
 		labelSvc:              labelService,
@@ -385,7 +368,7 @@ func (s *Service) processDocuments(ctx context.Context, resource Resource, webho
 			return err
 		}
 
-		bundlesFromDB, err := s.processBundles(ctx, resourceToAggregate.Type, resourceToAggregate.ID, doc.ConsumptionBundles, resourceHashes)
+		bundlesFromDB, err := s.bundleProcessor.Process(ctx, resourceToAggregate.Type, resourceToAggregate.ID, doc.ConsumptionBundles, resourceHashes)
 		if err != nil {
 			return err
 		}
@@ -679,222 +662,6 @@ func (s *Service) resyncApplicationTemplateVersionInTx(ctx context.Context, appT
 	return tx.Commit()
 }
 
-func (s *Service) processBundles(ctx context.Context, resourceType directorresource.Type, resourceID string, bundles []*model.BundleCreateInput, resourceHashes map[string]uint64) ([]*model.Bundle, error) {
-	bundlesFromDB, err := s.listBundlesInTx(ctx, resourceType, resourceID)
-	if err != nil {
-		return nil, err
-	}
-
-	credentialExchangeStrategyHashCurrent := uint64(0)
-	var credentialExchangeStrategyJSON gjson.Result
-	for _, bndl := range bundles {
-		bndlHash := resourceHashes[str.PtrStrToStr(bndl.OrdID)]
-		if err := s.resyncBundleInTx(ctx, resourceType, resourceID, bundlesFromDB, bndl, bndlHash); err != nil {
-			return nil, err
-		}
-
-		credentialExchangeStrategies, err := bndl.CredentialExchangeStrategies.MarshalJSON()
-		if err != nil {
-			return nil, errors.Wrapf(err, "while marshalling credential exchange strategies for %s with ID %s", resourceType, resourceID)
-		}
-
-		for _, credentialExchangeStrategy := range gjson.ParseBytes(credentialExchangeStrategies).Array() {
-			customType := credentialExchangeStrategy.Get(customTypeProperty).String()
-			isTenantMappingType := strings.Contains(customType, TenantMappingCustomTypeIdentifier)
-
-			if !isTenantMappingType {
-				continue
-			}
-
-			currentHash, err := HashObject(credentialExchangeStrategy)
-			if err != nil {
-				return nil, errors.Wrapf(err, "while hasing credential exchange strategy for application with ID %s", resourceID)
-			}
-
-			if credentialExchangeStrategyHashCurrent != 0 && currentHash != credentialExchangeStrategyHashCurrent {
-				return nil, errors.Errorf("There are differences in the Credential Exchange Strategies for Tenant Mappings for application with ID %s. They should be the same.", resourceID)
-			}
-
-			credentialExchangeStrategyHashCurrent = currentHash
-			credentialExchangeStrategyJSON = credentialExchangeStrategy
-		}
-	}
-
-	if err = s.resyncTenantMappingWebhooksInTx(ctx, credentialExchangeStrategyJSON, resourceID); err != nil {
-		return nil, err
-	}
-
-	bundlesFromDB, err = s.listBundlesInTx(ctx, resourceType, resourceID)
-	if err != nil {
-		return nil, err
-	}
-
-	return bundlesFromDB, nil
-}
-
-func (s *Service) listBundlesInTx(ctx context.Context, resourceType directorresource.Type, resourceID string) ([]*model.Bundle, error) {
-	tx, err := s.transact.Begin()
-	if err != nil {
-		return nil, err
-	}
-	defer s.transact.RollbackUnlessCommitted(ctx, tx)
-	ctx = persistence.SaveToContext(ctx, tx)
-
-	var bundlesFromDB []*model.Bundle
-	switch resourceType {
-	case directorresource.Application:
-		bundlesFromDB, err = s.bundleSvc.ListByApplicationIDNoPaging(ctx, resourceID)
-	case directorresource.ApplicationTemplateVersion:
-		bundlesFromDB, err = s.bundleSvc.ListByApplicationTemplateVersionIDNoPaging(ctx, resourceID)
-	}
-	if err != nil {
-		return nil, errors.Wrapf(err, "error while listing bundles for %s with id %q", resourceType, resourceID)
-	}
-
-	return bundlesFromDB, tx.Commit()
-}
-
-func (s *Service) resyncBundleInTx(ctx context.Context, resourceType directorresource.Type, resourceID string, bundlesFromDB []*model.Bundle, bundle *model.BundleCreateInput, bndlHash uint64) error {
-	tx, err := s.transact.Begin()
-	if err != nil {
-		return err
-	}
-	defer s.transact.RollbackUnlessCommitted(ctx, tx)
-	ctx = persistence.SaveToContext(ctx, tx)
-
-	if err := s.resyncBundle(ctx, resourceType, resourceID, bundlesFromDB, *bundle, bndlHash); err != nil {
-		return errors.Wrapf(err, "error while resyncing bundle with ORD ID %q", *bundle.OrdID)
-	}
-	return tx.Commit()
-}
-
-func (s *Service) resyncTenantMappingWebhooksInTx(ctx context.Context, credentialExchangeStrategyJSON gjson.Result, appID string) error {
-	if !credentialExchangeStrategyJSON.IsObject() {
-		log.C(ctx).Debugf("There are no tenant mappings to resync")
-		return nil
-	}
-
-	tenantMappingData, err := s.getTenantMappingData(credentialExchangeStrategyJSON, appID)
-	if err != nil {
-		return err
-	}
-
-	log.C(ctx).Infof("Enriching tenant mapping webhooks for application with ID %s", appID)
-
-	enrichedWebhooks, err := s.webhookSvc.EnrichWebhooksWithTenantMappingWebhooks([]*graphql.WebhookInput{createWebhookInput(credentialExchangeStrategyJSON, tenantMappingData)})
-	if err != nil {
-		return errors.Wrapf(err, "while enriching webhooks with tenant mapping webhooks for application with ID %s", appID)
-	}
-
-	ctxWithoutTenant := context.Background()
-	tx, err := s.transact.Begin()
-	if err != nil {
-		return err
-	}
-	defer s.transact.RollbackUnlessCommitted(ctxWithoutTenant, tx)
-
-	ctxWithoutTenant = persistence.SaveToContext(ctxWithoutTenant, tx)
-	ctxWithoutTenant = tenant.SaveToContext(ctxWithoutTenant, "", "")
-
-	appWebhooksFromDB, err := s.webhookSvc.ListForApplicationGlobal(ctxWithoutTenant, appID)
-	if err != nil {
-		return errors.Wrapf(err, "while listing webhooks from application with ID %s", appID)
-	}
-
-	tenantMappingRelatedWebhooksFromDB, enrichedWhModels, enrichedWhModelInputs, err := s.processEnrichedWebhooks(enrichedWebhooks, appWebhooksFromDB)
-	if err != nil {
-		return err
-	}
-
-	isEqual, err := isWebhookDataEqual(tenantMappingRelatedWebhooksFromDB, enrichedWhModels)
-	if err != nil {
-		return err
-	}
-
-	if isEqual {
-		log.C(ctxWithoutTenant).Infof("There are no differences in tenant mapping webhooks from the DB and the ORD document")
-		return tx.Commit()
-	}
-
-	log.C(ctxWithoutTenant).Infof("There are differences in tenant mapping webhooks from the DB and the ORD document. Continuing the sync.")
-
-	if err := s.deleteWebhooks(ctxWithoutTenant, tenantMappingRelatedWebhooksFromDB, appID); err != nil {
-		return err
-	}
-
-	if err := s.createWebhooks(ctxWithoutTenant, enrichedWhModelInputs, appID); err != nil {
-		return err
-	}
-
-	return tx.Commit()
-}
-
-func (s *Service) deleteWebhooks(ctx context.Context, webhooks []*model.Webhook, appID string) error {
-	for _, webhook := range webhooks {
-		log.C(ctx).Infof("Deleting webhook with ID %s for application %s", webhook.ID, appID)
-		if err := s.webhookSvc.Delete(ctx, webhook.ID, webhook.ObjectType); err != nil {
-			log.C(ctx).Errorf("error while deleting webhook with ID %s", webhook.ID)
-			return errors.Wrapf(err, "while deleting webhook with ID %s", webhook.ID)
-		}
-	}
-
-	return nil
-}
-
-func (s *Service) createWebhooks(ctx context.Context, webhooks []*model.WebhookInput, appID string) error {
-	for _, webhook := range webhooks {
-		log.C(ctx).Infof("Creating webhook with type %s for application %s", webhook.Type, appID)
-		if _, err := s.webhookSvc.Create(ctx, appID, *webhook, model.ApplicationWebhookReference); err != nil {
-			log.C(ctx).Errorf("error while creating webhook for app %s with type %s", appID, webhook.Type)
-			return errors.Wrapf(err, "error while creating webhook for app %s with type %s", appID, webhook.Type)
-		}
-	}
-
-	return nil
-}
-
-func (s *Service) getTenantMappingData(credentialExchangeStrategyJSON gjson.Result, appID string) (CredentialExchangeStrategyTenantMapping, error) {
-	tenantMappingType := credentialExchangeStrategyJSON.Get(customTypeProperty).String()
-	tenantMappingData, ok := s.config.credentialExchangeStrategyTenantMappings[tenantMappingType]
-	if !ok {
-		return CredentialExchangeStrategyTenantMapping{}, errors.Errorf("Credential Exchange Strategy has invalid %s value: %s for application with ID %s", customTypeProperty, tenantMappingType, appID)
-	}
-	return tenantMappingData, nil
-}
-
-func (s *Service) processEnrichedWebhooks(enrichedWebhooks []*graphql.WebhookInput, webhooksFromDB []*model.Webhook) ([]*model.Webhook, []*model.Webhook, []*model.WebhookInput, error) {
-	tenantMappingRelatedWebhooksFromDB := make([]*model.Webhook, 0)
-	enrichedWebhookModels := make([]*model.Webhook, 0)
-	enrichedWebhookModelInputs := make([]*model.WebhookInput, 0)
-
-	for _, wh := range enrichedWebhooks {
-		convertedIn, err := s.webhookConverter.InputFromGraphQL(wh)
-		if err != nil {
-			return nil, nil, nil, errors.Wrap(err, "while converting the WebhookInput")
-		}
-
-		enrichedWebhookModelInputs = append(enrichedWebhookModelInputs, convertedIn)
-
-		webhookModel := convertedIn.ToWebhook("", "", "")
-
-		for _, webhookFromDB := range webhooksFromDB {
-			if webhookFromDB.Type == convertedIn.Type {
-				webhookModel.ID = webhookFromDB.ID
-				webhookModel.ObjectType = webhookFromDB.ObjectType
-				webhookModel.ObjectID = webhookFromDB.ObjectID
-				webhookModel.CreatedAt = webhookFromDB.CreatedAt
-
-				tenantMappingRelatedWebhooksFromDB = append(tenantMappingRelatedWebhooksFromDB, webhookFromDB)
-				break
-			}
-		}
-
-		enrichedWebhookModels = append(enrichedWebhookModels, webhookModel)
-	}
-
-	return tenantMappingRelatedWebhooksFromDB, enrichedWebhookModels, enrichedWebhookModelInputs, nil
-}
-
 func (s *Service) processAPIs(ctx context.Context, resourceType directorresource.Type, resourceID string, bundlesFromDB []*model.Bundle, packagesFromDB []*model.Package, apis []*model.APIDefinitionInput, resourceHashes map[string]uint64) ([]*model.APIDefinition, []*ordFetchRequest, error) {
 	apisFromDB, err := s.listAPIsInTx(ctx, resourceType, resourceID)
 	if err != nil {
@@ -1026,18 +793,6 @@ func (s *Service) resyncEventInTx(ctx context.Context, resourceType directorreso
 		return nil, errors.Wrapf(err, "error while resyncing event with ORD ID %q", *event.OrdID)
 	}
 	return fetchRequests, tx.Commit()
-}
-
-func (s *Service) resyncBundle(ctx context.Context, resourceType directorresource.Type, resourceID string, bundlesFromDB []*model.Bundle, bndl model.BundleCreateInput, bndlHash uint64) error {
-	ctx = addFieldToLogger(ctx, "bundle_ord_id", *bndl.OrdID)
-	if i, found := searchInSlice(len(bundlesFromDB), func(i int) bool {
-		return equalStrings(bundlesFromDB[i].OrdID, bndl.OrdID)
-	}); found {
-		return s.bundleSvc.UpdateBundle(ctx, resourceType, bundlesFromDB[i].ID, bundleUpdateInputFromCreateInput(bndl), bndlHash)
-	}
-
-	_, err := s.bundleSvc.CreateBundle(ctx, resourceType, resourceID, bndl, bndlHash)
-	return err
 }
 
 func (s *Service) resyncAppTemplateVersion(ctx context.Context, appTemplateID string, appTemplateVersionsFromDB []*model.ApplicationTemplateVersion, appTemplateVersion *model.ApplicationTemplateVersionInput) error {
@@ -1793,22 +1548,6 @@ func hashResources(docs Documents) (map[string]uint64, error) {
 	return resourceHashes, nil
 }
 
-func bundleUpdateInputFromCreateInput(in model.BundleCreateInput) model.BundleUpdateInput {
-	return model.BundleUpdateInput{
-		Name:                           in.Name,
-		Description:                    in.Description,
-		InstanceAuthRequestInputSchema: in.InstanceAuthRequestInputSchema,
-		DefaultInstanceAuth:            in.DefaultInstanceAuth,
-		OrdID:                          in.OrdID,
-		ShortDescription:               in.ShortDescription,
-		Links:                          in.Links,
-		Labels:                         in.Labels,
-		DocumentationLabels:            in.DocumentationLabels,
-		CredentialExchangeStrategies:   in.CredentialExchangeStrategies,
-		CorrelationIDs:                 in.CorrelationIDs,
-	}
-}
-
 // extractDefaultConsumptionBundle converts the defaultConsumptionBundle which is bundle ORD_ID into internal bundle_id
 func extractDefaultConsumptionBundle(bundlesFromDB []*model.Bundle, defaultConsumptionBundle *string) string {
 	var bundleID string
@@ -1884,46 +1623,6 @@ func addFieldToLogger(ctx context.Context, fieldName, fieldValue string) context
 	logger := log.LoggerFromContext(ctx)
 	logger = logger.WithField(fieldName, fieldValue)
 	return log.ContextWithLogger(ctx, logger)
-}
-
-func createWebhookInput(credentialExchangeStrategyJSON gjson.Result, tenantMappingData CredentialExchangeStrategyTenantMapping) *graphql.WebhookInput {
-	inputMode := graphql.WebhookMode(tenantMappingData.Mode)
-	return &graphql.WebhookInput{
-		URL: str.Ptr(credentialExchangeStrategyJSON.Get(callbackURLProperty).String()),
-		Auth: &graphql.AuthInput{
-			AccessStrategy: str.Ptr(string(accessstrategy.CMPmTLSAccessStrategy)),
-		},
-		Mode:    &inputMode,
-		Version: str.Ptr(tenantMappingData.Version),
-	}
-}
-
-func isWebhookDataEqual(tenantMappingRelatedWebhooksFromDB, enrichedWhModels []*model.Webhook) (bool, error) {
-	appWhsFromDBMarshaled, err := json.Marshal(tenantMappingRelatedWebhooksFromDB)
-	if err != nil {
-		return false, errors.Wrapf(err, "while marshalling webhooks from DB")
-	}
-
-	appWhsFromDBHash, err := HashObject(string(appWhsFromDBMarshaled))
-	if err != nil {
-		return false, errors.Wrapf(err, "while hashing webhooks from DB")
-	}
-
-	enrichedWhsMarshaled, err := json.Marshal(enrichedWhModels)
-	if err != nil {
-		return false, errors.Wrapf(err, "while marshalling webhooks from DB")
-	}
-
-	enrichedHash, err := HashObject(string(enrichedWhsMarshaled))
-	if err != nil {
-		return false, errors.Wrapf(err, "while hashing webhooks from ORD document")
-	}
-
-	if strconv.FormatUint(appWhsFromDBHash, 10) == strconv.FormatUint(enrichedHash, 10) {
-		return true, nil
-	}
-
-	return false, nil
 }
 
 func (s *Service) saveLowestOwnerForAppToContextInTx(ctx context.Context, appID string) (context.Context, error) {
